// angularTutorialEmployeeSystem/src/app/model/Client.ts
export interface Client {
    clientId:          number;
    contactPersonName: string;
    companyName:       string;
    address:           string;
    city:              string;
    pincode:           string;
    state:             string;
    EmployeeStrength:  number;
    gstNo:             string;
    contactNo:         string;
    regNo:             string;
}